fixed build

+++ miner.h
@@ -357,7 +357,7 @@ extern int scanhash_s3(int thr_id, uint32_t *pdata,
 
 extern int scanhash_whc(int thr_id, uint32_t *pdata,
 	uint32_t *ptarget, uint32_t max_nonce,
-	uint32_t *hashes_done);
+	unsigned long *hashes_done);
 
 extern int scanhash_whirlpoolx(int thr_id, uint32_t *pdata,
 	uint32_t *ptarget, uint32_t max_nonce,
